requirements.txt
Flask
requests
SQLAlchemy
Werkzeug
Flask_Login
Flask_SqlAlchemy
Flask_Migrate
Flask_Restful
Flask_Cors
PyJWT
pandas
numpy
matplotlib
seaborn
scikit-learn
openai
langchain
streamlit
google-generativeai